:lock: Improve workspace API role filtering

+++ apps/builder/src/features/workspace/api/deleteWorkspace.ts
@@ -1,6 +1,8 @@
 import prisma from '@/lib/prisma'
 import { authenticatedProcedure } from '@/helpers/server/trpc'
 import { z } from 'zod'
+import { isAdminWriteWorkspaceForbidden } from '../helpers/isAdminWriteWorkspaceForbidden'
+import { TRPCError } from '@trpc/server'
 
 export const deleteWorkspace = authenticatedProcedure
   .meta({
@@ -23,6 +25,14 @@ export const deleteWorkspace = authenticatedProcedure
     })
   )
   .mutation(async ({ input: { workspaceId }, ctx: { user } }) => {
+    const workspace = await prisma.workspace.findFirst({
+      where: { id: workspaceId },
+      include: { members: true },
+    })
+
+    if (!workspace || (await isAdminWriteWorkspaceForbidden(workspace, user)))
+      throw new TRPCError({ code: 'NOT_FOUND', message: 'No workspaces found' })
+
     await prisma.workspace.deleteMany({
       where: { members: { some: { userId: user.id } }, id: workspaceId },
     })
